% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoREADME.R
\name{autoREADME}
\alias{autoREADME}
\title{Auto generate or update README}
\usage{
autoREADME(dirREADME = NULL, title = "", description = "")
}
\arguments{
\item{dirREADME}{A string specifying the directory to auto generate or update a README file, no default. If updating an existing file, the README must already contain a File and Folder table preceeded by a table.}

\item{title}{An optional argument (string) to provide a new title when generating a README file for the first time, default = "".}

\item{description}{An optional argument (string) to provide a description to follow the title when generating a new README file, default = "".}
}
\value{
Generates new/updated README file in specified dirREADME directory.
}
\description{
Auto generate or update a README file in the specified directory.
}
