% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_regex_match_display.R
\name{compute_regex_match_display}
\alias{compute_regex_match_display}
\title{Compute which parts of a string are matched by a regex}
\usage{
compute_regex_match_display(str, pattern)
}
\arguments{
\item{str}{string which regex will be matched against}

\item{pattern}{string containing the regex pattern}
}
\value{
a string of underscore characters in places matched by the regex
(and whitespace for non-matching characters)
}
\description{
Compute which parts of a string are matched by a regex. Used by the
str_print_regex_matches() function.
}
\examples{
str_print_regex_matches("hello world!", "l.*r")
#> "  _______   "

str_print_regex_matches("hello world!", "[a-z]")
#> "_____ _____ "

}
