% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_num_to_str.R
\name{date_num_to_str}
\alias{date_num_to_str}
\title{Convert number representing day/month into string equivalent}
\usage{
date_num_to_str(x, num_to, lower = FALSE, longform = FALSE)
}
\arguments{
\item{x}{number representing a day or a month}

\item{num_to}{string specifying what type of date to convert ("day" or
"month")}

\item{lower}{boolean specifying whether to return lower case string (default
is to capitalise the first character)}

\item{longform}{boolean specifying whether to return the entire string
(default is to abbreviate to the first 3 letters)}
}
\value{
a string representing the day or month
}
\description{
Convert a number describing a day or a month into its string equivalent.
Can give an abbreviated or long form string
}
\examples{
date_num_to_str(1, num_to = "day")
#> "Sun"

date_num_to_str(2, num_to = "day", lower = TRUE, longform = TRUE)
#> "monday"

date_num_to_str(12, num_to = "month")
#> "Dec"

}
