% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{date_str_to_num}
\alias{date_str_to_num}
\title{Convert string abbreviation of day/month into numeric equivalent}
\usage{
date_str_to_num(x)
}
\arguments{
\item{x}{string A three letter string representing a day or a month}
}
\value{
a number
}
\description{
Convert a string abbreviation of a day or a month into its numeric
equivalent
}
\examples{
date_str_to_num("Sun")

#> 1

date_str_to_num("Mon")

#> 2


date_str_to_num("Dec")

#> 12

}
