% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_showmatches.R
\name{str_showmatches}
\alias{str_showmatches}
\title{Visualise which parts of a string are matched by a regex}
\usage{
str_showmatches(str, pattern)
}
\arguments{
\item{str}{string which regex will be matched against}

\item{pattern}{string containing the regex pattern}
}
\value{
prints the regex pattern, the original string where matches are
indicated by a horizontal line above
}
\description{
Visualise which parts of a string are matched by a regex
}
\examples{
str_showmatches("hello world!", "l.*r")
#> pattern:  l.*r
#> matches:    _______
#> string :  hello world!

str_showmatches("hello world!", "[a-z]")
#> pattern:  [a-z]
#> matches:  _____ _____
#> string :  hello world!

}
