% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_str_to_num.R
\name{date_str_to_num}
\alias{date_str_to_num}
\title{Convert string representing day/month into numeric equivalent}
\usage{
date_str_to_num(x)
}
\arguments{
\item{x}{string representing a day or a month}
}
\value{
a number representing the day or month
}
\description{
Convert a string describing a day or a month into its numeric equivalent.
Works with abbreviations and is case insensitive.
}
\examples{
date_str_to_num("Sun")
#> 1

date_str_to_num("monday")
#> 2

date_str_to_num("Dec")
#> 12

}
